feat: adição do parametro id e readonly nos inputs, finalização da pá…

+++ front-end/src/mobile/pages/login/Login.jsx
@@ -6,38 +6,74 @@ import styles from "./Login.module.css";
 import InputFormulario from "../../components/inputs/inputFormularios/InputFormulario";
 import Botao from "../../components/botoes/botaoLoginCadastro/Botao";
 import Header from "../../components/header/Header";
+import { ToastContainer } from "react-toastify";
 
 const Login = () => {
+
+    sessionStorage.nome = "";
+    sessionStorage.situacao = "";
+    sessionStorage.userId = "";
+    sessionStorage.token = "";
+
     const navigate = useNavigate();
 
-    const[email, setEmail] = useState("");
-    const[senha, setSenha] = useState("");
+    const [email, setEmail] = useState("");
+    const [senha, setSenha] = useState("");
+    const [showConfirmation, setShowConfirmation] = useState(false);
 
-    const handleInputChange = (event, setStateFunction) => { // Função para manipular as mudanças nos inputs
-        setStateFunction(event.target.value); // Atualiza o estado correspondente
+    const handleInputChange = (event, setStateFunction) => {
+        setStateFunction(event.target.value);
     }
-    
-    const handleClick = () => {
-        const usuario = { // Cria um objeto com os dados do formulário
+
+    const handleClick = async () => {
+        const usuario = {
             email,
             senha
         };
-        
-        api.post(`/usuarios/login`, usuario).then(response => {
+
+        try {
+            const response = await api.post(`/usuarios/login`, usuario);
             const sessionData = response.data.token;
             const sessionUserId = response.data.userId;
             const sessionUserName = response.data.nome;
 
-            sessionStorage.setItem('token', sessionData)
-            sessionStorage.setItem('userId', sessionUserId)
-            sessionStorage.setItem('nome', sessionUserName)
-            
-            toast.success("Login Efetuado com sucesso!"); // Exibe uma mensagem de sucesso
-            navigate("/feedGeral"); // Redireciona para a página de músicas
-        }).catch(() => {
-            toast.error("Ocorreu um erro ao salvar os dados, por favor, tente novamente."); // Exibe uma mensagem de erro se a requisição falhar
-        });
+            const situacaoUsuario = await buscarSituacaoUser(sessionUserId);
+
+            if (situacaoUsuario === 'PENDENTE' || situacaoUsuario === 'NEGADO') {
+                setShowConfirmation(true);
+            } else {
+                sessionStorage.setItem('token', sessionData);
+                sessionStorage.setItem('userId', sessionUserId);
+                sessionStorage.setItem('nome', sessionUserName);
+
+                if(endereco === null){
+                    toast.success("Login Efetuado com sucesso!");
+                    navigate("/completarCadastro");
+                } else{
+                    toast.success("Login Efetuado com sucesso!");
+                    navigate("/feedGeral");
+                }
+            }
+        } catch (error) {
+            toast.error("Ocorreu um erro ao salvar os dados, por favor, tente novamente.");
+        }
     };
+
+    var endereco;
+
+    const buscarSituacaoUser = async (id) => {
+        try {
+            const response = await api.get(`/usuarios/${id}`);
+            const situacao = response.data.tipoStatus.status;
+            sessionStorage.setItem('situacao', situacao);
+            console.log(response.data);
+            endereco = response.data.endereco;
+            return situacao;
+        } catch (error) {
+            console.error("Erro ao buscar situação do usuário:", error);
+            return null; // ou algum valor padrão
+        }
+    }
 
     return (
         <>
@@ -46,8 +82,8 @@ const Login = () => {
                 <h1 className={styles['loginMobileTitulo']}>Login</h1>
                 <h3>Bem-Vindo de volta</h3>
                 <div className={styles['divInputs']}>
-                    <InputFormulario  placeHolder="Email" value={email} onChange={(e) => handleInputChange(e, setEmail)}/>
-                    <InputFormulario placeHolder="Senha" value={senha} onChange={(e) => handleInputChange(e, setSenha)}/>
+                    <InputFormulario placeHolder="Email" value={email} onChange={(e) => handleInputChange(e, setEmail)} />
+                    <InputFormulario placeHolder="Senha" value={senha} onChange={(e) => handleInputChange(e, setSenha)} />
                 </div>
                 <a href=""><h2>Esqueceu sua senha?</h2></a>
                 <div className={styles['divBotao']}>
@@ -55,6 +91,29 @@ const Login = () => {
                 </div>
 
             </div>
+
+            {showConfirmation && (
+                <div className={styles['modalOverlay']}>
+                    <div className={styles['modalContent']}>
+                        <h3>Que pena!</h3>
+                        <p>Você ainda não possui acesso a plataforma, entre em contato com o instituto.</p>
+                        <button className={styles['confirmButton']} onClick={() => {setShowConfirmation(false); window.location.reload();}}>Ok</button>
+                    </div>
+                </div>
+            )}
+
+            <ToastContainer
+                position="top-right"
+                autoClose={1000}
+                hideProgressBar={false}
+                newestOnTop={false}
+                closeOnClick
+                rtl={false}
+                pauseOnFocusLoss
+                draggable
+                pauseOnHover
+                theme="colored"
+            />
         </>
     )
 };
